feat: add regular navbar implementation for rest of srcbook views
